Backend/requirements2.txt
llama-cpp-python
cached_property
farm-haystack
fastapi
uvicorn
farm-haystack[weaviate]
pypdf
torch
sentence_transformers
haystack-ai